+++ projects-ms/src/project/router.rs
@@ -5,6 +5,7 @@ use crate::user::AccessTokenClaims;
 use crate::AppState;
 use axum::extract::{Path, State};
 use axum::http::StatusCode;
+use axum::response::IntoResponse;
 use axum::routing::get;
 use axum::{debug_handler, Json, Router};
 use serde::Deserialize;
@@ -47,16 +48,26 @@ async fn get_projects(
 #[debug_handler]
 async fn get_project(
     Path(project_id): Path<Uuid>,
+    user: AccessTokenClaims,
     State(state): State<AppState>,
-) -> Result<Json<Project>> {
-    controller::get_project(project_id, state).await.map(Json)
+) -> Result<impl IntoResponse> {
+    let project = controller::get_project(project_id, state).await?;
+    if project.user_id != user.sub {
+        return Ok(StatusCode::FORBIDDEN.into_response());
+    }
+    Ok(Json(project).into_response())
 }
 
 #[debug_handler]
 async fn delete_project(
     Path(project_id): Path<Uuid>,
+    user: AccessTokenClaims,
     State(state): State<AppState>,
-) -> Result<StatusCode> {
+) -> Result<impl IntoResponse> {
+    let project = controller::get_project(project_id, state.clone()).await?;
+    if project.user_id != user.sub {
+        return Ok(StatusCode::FORBIDDEN.into_response());
+    }
     controller::delete_project(project_id, state).await?;
-    Ok(StatusCode::NO_CONTENT)
+    Ok(Json(project).into_response())
 }
